requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
mysql-connector-python==8.3.0
PyMySQL==1.1.0
python-dotenv==1.0.1
Werkzeug==3.0.1
email-validator==2.1.0.post1 